feat: add WordsList component file
